lots more refactoring
