+++ py/legacyhalos/html.py
@@ -789,334 +789,3 @@ def viewer_link(ra, dec, width, sga=False, manga=False):
 
     return viewer
 
-def skyserver_link(gal):
-    if 'SDSS_OBJID' in gal.colnames:
-        return 'http://skyserver.sdss.org/dr14/en/tools/explore/summary.aspx?id={:d}'.format(gal['SDSS_OBJID'])
-    else:
-        return ''
-
-def make_html(sample=None, datadir=None, htmldir=None, bands=('g', 'r', 'z'),
-              refband='r', pixscale=0.262, zcolumn='Z', intflux=None,
-              first=None, last=None, nproc=1, survey=None, makeplots=True,
-              clobber=False, verbose=True, maketrends=False, ccdqa=False):
-    """Make the HTML pages.
-
-    """
-    import subprocess
-    import fitsio
-    from legacyhalos.misc import RADIUS_CLUSTER_KPC as radius_mosaic_kpc
-
-    if datadir is None:
-        datadir = legacyhalos.io.legacyhalos_data_dir()
-    if htmldir is None:
-        htmldir = legacyhalos.io.legacyhalos_html_dir()
-
-    if sample is None:
-        sample = legacyhalos.io.read_sample(first=first, last=last)
-
-    if type(sample) is astropy.table.row.Row:
-        sample = astropy.table.Table(sample)
-
-    if zcolumn is None:
-        zcolumn = ZCOLUMN
-        
-    galaxy, galaxydir, htmlgalaxydir = legacyhalos.io.get_galaxy_galaxydir(sample, html=True)
-
-    # Write the last-updated date to a webpage.
-    js = html_javadate()       
-
-    # Get the viewer link
-    def _viewer_link(gal):
-        baseurl = 'http://legacysurvey.org/viewer/'
-        width = 2 * legacyhalos.misc.cutout_radius_kpc(
-            redshift=gal[zcolumn], pixscale=0.262,
-            radius_kpc=radius_mosaic_kpc) # [pixels]
-        if width > 400:
-            zoom = 14
-        else:
-            zoom = 15
-        viewer = '{}?ra={:.6f}&dec={:.6f}&zoom={:g}&layer=dr8'.format(
-            baseurl, gal['RA'], gal['DEC'], zoom)
-        
-        return viewer
-
-    def _skyserver_link(gal):
-        if 'SDSS_OBJID' in gal.colnames:
-            return 'http://skyserver.sdss.org/dr14/en/tools/explore/summary.aspx?id={:d}'.format(gal['SDSS_OBJID'])
-        else:
-            return ''
-
-    trendshtml = 'trends.html'
-    homehtml = 'index.html'
-
-    # Build the home (index.html) page--
-    if not os.path.exists(htmldir):
-        os.makedirs(htmldir)
-        #shutil.chown(htmldir, group='cosmo')
-    homehtmlfile = os.path.join(htmldir, homehtml)
-
-    #if verbose:
-    print('Writing {}'.format(homehtmlfile))
-    with open(homehtmlfile, 'w') as html:
-        html.write('<html><body>\n')
-        html.write('<style type="text/css">\n')
-        html.write('table, td, th {padding: 5px; text-align: left; border: 1px solid black;}\n')
-        html.write('</style>\n')
-
-        html.write('<h1>LegacyHalos: Central Galaxies</h1>\n')
-        if maketrends:
-            html.write('<p>\n')
-            html.write('<a href="{}">Sample Trends</a><br />\n'.format(trendshtml))
-            html.write('<a href="https://github.com/moustakas/legacyhalos">Code and documentation</a>\n')
-            html.write('</p>\n')
-
-        html.write('<table>\n')
-        html.write('<tr>\n')
-        html.write('<th>Number</th>\n')
-        html.write('<th>Galaxy</th>\n')
-        html.write('<th>RA</th>\n')
-        html.write('<th>Dec</th>\n')
-        html.write('<th>Redshift</th>\n')
-        html.write('<th>Richness</th>\n')
-        html.write('<th>Pcen</th>\n')
-        html.write('<th>Viewer</th>\n')
-        html.write('<th>SkyServer</th>\n')
-        html.write('</tr>\n')
-        for ii, (gal, galaxy1, htmlgalaxydir1) in enumerate(zip(
-            sample, np.atleast_1d(galaxy), np.atleast_1d(htmlgalaxydir) )):
-
-            htmlfile1 = os.path.join(htmlgalaxydir1.replace(htmldir, '')[1:], '{}.html'.format(galaxy1))
-
-            html.write('<tr>\n')
-            html.write('<td>{:g}</td>\n'.format(ii))
-            html.write('<td><a href="{}">{}</a></td>\n'.format(htmlfile1, galaxy1))
-            html.write('<td>{:.7f}</td>\n'.format(gal['RA']))
-            html.write('<td>{:.7f}</td>\n'.format(gal['DEC']))
-            html.write('<td>{:.5f}</td>\n'.format(gal[zcolumn]))
-            html.write('<td>{:.4f}</td>\n'.format(gal['LAMBDA_CHISQ']))
-            html.write('<td>{:.3f}</td>\n'.format(gal['P_CEN'][0]))
-            html.write('<td><a href="{}" target="_blank">Link</a></td>\n'.format(_viewer_link(gal)))
-            html.write('<td><a href="{}" target="_blank">Link</a></td>\n'.format(_skyserver_link(gal)))
-            html.write('</tr>\n')
-        html.write('</table>\n')
-        
-        html.write('<br /><br />\n')
-        html.write('<b><i>Last updated {}</b></i>\n'.format(js))
-        html.write('</html></body>\n')
-        html.close()
-
-    # Build the trends (trends.html) page--
-    if maketrends:
-        trendshtmlfile = os.path.join(htmldir, trendshtml)
-        if verbose:
-            print('Writing {}'.format(trendshtmlfile))
-        with open(trendshtmlfile, 'w') as html:
-            html.write('<html><body>\n')
-            html.write('<style type="text/css">\n')
-            html.write('table, td, th {padding: 5px; text-align: left; border: 1px solid black;}\n')
-            html.write('</style>\n')
-
-            html.write('<h1>LegacyHalos: Sample Trends</h1>\n')
-            html.write('<p><a href="https://github.com/moustakas/legacyhalos">Code and documentation</a></p>\n')
-            html.write('<a href="trends/ellipticity_vs_sma.png"><img src="trends/ellipticity_vs_sma.png" alt="Missing file ellipticity_vs_sma.png" height="auto" width="50%"></a>')
-            html.write('<a href="trends/gr_vs_sma.png"><img src="trends/gr_vs_sma.png" alt="Missing file gr_vs_sma.png" height="auto" width="50%"></a>')
-            html.write('<a href="trends/rz_vs_sma.png"><img src="trends/rz_vs_sma.png" alt="Missing file rz_vs_sma.png" height="auto" width="50%"></a>')
-
-            html.write('<br /><br />\n')
-            html.write('<b><i>Last updated {}</b></i>\n'.format(js))
-            html.write('</html></body>\n')
-            html.close()
-
-    nextgalaxy = np.roll(np.atleast_1d(galaxy), -1)
-    prevgalaxy = np.roll(np.atleast_1d(galaxy), 1)
-    nexthtmlgalaxydir = np.roll(np.atleast_1d(htmlgalaxydir), -1)
-    prevhtmlgalaxydir = np.roll(np.atleast_1d(htmlgalaxydir), 1)
-
-    # Make a separate HTML page for each object.
-    for ii, (gal, galaxy1, galaxydir1, htmlgalaxydir1) in enumerate( zip(
-        sample, np.atleast_1d(galaxy), np.atleast_1d(galaxydir), np.atleast_1d(htmlgalaxydir) ) ):
-
-        radius_mosaic_arcsec = legacyhalos.misc.cutout_radius_kpc(
-            redshift=gal[zcolumn], radius_kpc=radius_mosaic_kpc) # [arcsec]
-        diam_mosaic_pixels = _mosaic_width(radius_mosaic_arcsec, pixscale)
-
-        ellipse = legacyhalos.io.read_ellipsefit(galaxy1, galaxydir1, verbose=verbose)
-        
-        if not os.path.exists(htmlgalaxydir1):
-            os.makedirs(htmlgalaxydir1, mode=0o775)
-            #shutil.chown(htmlgalaxydir1, group='cosmo')
-
-        ccdsfile = os.path.join(galaxydir1, '{}-ccds.fits'.format(galaxy1))
-        if os.path.isfile(ccdsfile):
-            nccds = fitsio.FITS(ccdsfile)[1].get_nrows()
-        else:
-            nccds = None
-
-        nexthtmlgalaxydir1 = os.path.join('{}'.format(nexthtmlgalaxydir[ii].replace(htmldir, '')[1:]), '{}.html'.format(nextgalaxy[ii]))
-        prevhtmlgalaxydir1 = os.path.join('{}'.format(prevhtmlgalaxydir[ii].replace(htmldir, '')[1:]), '{}.html'.format(prevgalaxy[ii]))
-
-        htmlfile = os.path.join(htmlgalaxydir1, '{}.html'.format(galaxy1))
-        with open(htmlfile, 'w') as html:
-            html.write('<html><body>\n')
-            html.write('<style type="text/css">\n')
-            html.write('table, td, th {padding: 5px; text-align: left; border: 1px solid black;}\n')
-            html.write('</style>\n')
-
-            html.write('<h1>Central Galaxy {}</h1>\n'.format(galaxy1))
-
-            html.write('<a href="../../../../{}">Home</a>\n'.format(homehtml))
-            html.write('<br />\n')
-            html.write('<a href="../../../../{}">Next Central Galaxy ({})</a>\n'.format(nexthtmlgalaxydir1, nextgalaxy[ii]))
-            html.write('<br />\n')
-            html.write('<a href="../../../../{}">Previous Central Galaxy ({})</a>\n'.format(prevhtmlgalaxydir1, prevgalaxy[ii]))
-            html.write('<br />\n')
-            html.write('<br />\n')
-
-            # Table of properties
-            html.write('<table>\n')
-            html.write('<tr>\n')
-            html.write('<th>Number</th>\n')
-            html.write('<th>Galaxy</th>\n')
-            html.write('<th>RA</th>\n')
-            html.write('<th>Dec</th>\n')
-            html.write('<th>Redshift</th>\n')
-            html.write('<th>Richness</th>\n')
-            html.write('<th>Pcen</th>\n')
-            html.write('<th>Viewer</th>\n')
-            html.write('<th>SkyServer</th>\n')
-            html.write('</tr>\n')
-
-            html.write('<tr>\n')
-            html.write('<td>{:g}</td>\n'.format(ii))
-            html.write('<td>{}</td>\n'.format(galaxy1))
-            html.write('<td>{:.7f}</td>\n'.format(gal['RA']))
-            html.write('<td>{:.7f}</td>\n'.format(gal['DEC']))
-            html.write('<td>{:.5f}</td>\n'.format(gal[zcolumn]))
-            html.write('<td>{:.4f}</td>\n'.format(gal['LAMBDA_CHISQ']))
-            html.write('<td>{:.3f}</td>\n'.format(gal['P_CEN'][0]))
-            html.write('<td><a href="{}" target="_blank">Link</a></td>\n'.format(_viewer_link(gal)))
-            html.write('<td><a href="{}" target="_blank">Link</a></td>\n'.format(_skyserver_link(gal)))
-            html.write('</tr>\n')
-            html.write('</table>\n')
-
-            html.write('<h2>Image Mosaics</h2>\n')
-            html.write('<p>Each mosaic (left to right: data, model of all but the central galaxy, and residual image containing just the central galaxy) is {:.0f} kpc = {:.3f} arcsec = {:.0f} pixels in diameter.</p>\n'.format(2*radius_mosaic_kpc, 2*radius_mosaic_arcsec, diam_mosaic_pixels))
-            #html.write('<br />\n')
-
-            html.write('<table width="90%">\n')
-            html.write('<tr><td><a href="{}-grz-montage.png"><img src="{}-grz-montage.png" alt="Missing file {}-grz-montage.png" height="auto" width="100%"></a></td></tr>\n'.format(galaxy1, galaxy1, galaxy1))
-            #html.write('<tr><td>Data, Model, Residuals</td></tr>\n')
-            html.write('</table>\n')
-            #html.write('<br />\n')
-            
-            html.write('<h2>Elliptical Isophote Analysis</h2>\n')
-            html.write('<table width="90%">\n')
-            html.write('<tr>\n')
-            pngfile = '{}-ellipse-multiband.png'.format(galaxy1)
-            html.write('<td><a href="{0}"><img src="{0}" alt="Missing file {0}" height="auto" width="100%"></a></td>\n'.format(pngfile))
-            html.write('</tr>\n')
-            html.write('</table>\n')
-
-            html.write('<table width="90%">\n')
-            html.write('<tr>\n')
-            #html.write('<td><a href="{}-ellipse-ellipsefit.png"><img src="{}-ellipse-ellipsefit.png" alt="Missing file {}-ellipse-ellipsefit.png" height="auto" width="100%"></a></td>\n'.format(galaxy1, galaxy1, galaxy1))
-            pngfile = '{}-ellipse-sbprofile.png'.format(galaxy1)
-            html.write('<td width="50%"><a href="{0}"><img src="{0}" alt="Missing file {0}" height="auto" width="100%"></a></td>\n'.format(pngfile))
-            pngfile = '{}-ellipse-cog.png'.format(galaxy1)
-            html.write('<td><a href="{0}"><img src="{0}" alt="Missing file {0}" height="auto" width="100%"></a></td>\n'.format(pngfile))
-            html.write('<td></td>\n')
-            html.write('</tr>\n')
-            html.write('</table>\n')
-
-            if False:
-                html.write('<h2>Surface Brightness Profile Modeling</h2>\n')
-                html.write('<table width="90%">\n')
-
-                # single-sersic
-                html.write('<tr>\n')
-                html.write('<th>Single Sersic (No Wavelength Dependence)</th><th>Single Sersic</th>\n')
-                html.write('</tr>\n')
-                html.write('<tr>\n')
-                html.write('<td><a href="{}-sersic-single-nowavepower.png"><img src="{}-sersic-single-nowavepower.png" alt="Missing file {}-sersic-single-nowavepower.png" height="auto" width="100%"></a></td>\n'.format(galaxy1, galaxy1, galaxy1))
-                html.write('<td><a href="{}-sersic-single.png"><img src="{}-sersic-single.png" alt="Missing file {}-sersic-single.png" height="auto" width="100%"></a></td>\n'.format(galaxy1, galaxy1, galaxy1))
-                html.write('</tr>\n')
-
-                # Sersic+exponential
-                html.write('<tr>\n')
-                html.write('<th>Sersic+Exponential (No Wavelength Dependence)</th><th>Sersic+Exponential</th>\n')
-                html.write('</tr>\n')
-                html.write('<tr>\n')
-                html.write('<td><a href="{}-sersic-exponential-nowavepower.png"><img src="{}-sersic-exponential-nowavepower.png" alt="Missing file {}-sersic-exponential-nowavepower.png" height="auto" width="100%"></a></td>\n'.format(galaxy1, galaxy1, galaxy1))
-                html.write('<td><a href="{}-sersic-exponential.png"><img src="{}-sersic-exponential.png" alt="Missing file {}-sersic-exponential.png" height="auto" width="100%"></a></td>\n'.format(galaxy1, galaxy1, galaxy1))
-                html.write('</tr>\n')
-
-                # double-sersic
-                html.write('<tr>\n')
-                html.write('<th>Double Sersic (No Wavelength Dependence)</th><th>Double Sersic</th>\n')
-                html.write('</tr>\n')
-                html.write('<tr>\n')
-                html.write('<td><a href="{}-sersic-double-nowavepower.png"><img src="{}-sersic-double-nowavepower.png" alt="Missing file {}-sersic-double-nowavepower.png" height="auto" width="100%"></a></td>\n'.format(galaxy1, galaxy1, galaxy1))
-                html.write('<td><a href="{}-sersic-double.png"><img src="{}-sersic-double.png" alt="Missing file {}-sersic-double.png" height="auto" width="100%"></a></td>\n'.format(galaxy1, galaxy1, galaxy1))
-                html.write('</tr>\n')
-
-                html.write('</table>\n')
-
-                html.write('<br />\n')
-
-            if nccds and ccdqa:
-                html.write('<h2>CCD Diagnostics</h2>\n')
-                html.write('<table width="90%">\n')
-                html.write('<tr>\n')
-                pngfile = '{}-ccdpos.png'.format(galaxy1)
-                html.write('<td><a href="{0}"><img src="{0}" alt="Missing file {0}" height="auto" width="100%"></a></td>\n'.format(
-                    pngfile))
-                html.write('</tr>\n')
-
-                for iccd in range(nccds):
-                    html.write('<tr>\n')
-                    pngfile = '{}-2d-ccd{:02d}.png'.format(galaxy1, iccd)
-                    html.write('<td><a href="{0}"><img src="{0}" alt="Missing file {0}" height="auto" width="100%"></a></td>\n'.format(
-                        pngfile))
-                    html.write('</tr>\n')
-                html.write('</table>\n')
-                html.write('<br />\n')
-
-            html.write('<a href="../../../../{}">Home</a>\n'.format(homehtml))
-            html.write('<br />\n')
-            html.write('<a href="{}">Next Central Galaxy ({})</a>\n'.format(nexthtmlgalaxydir1, nextgalaxy[ii]))
-            html.write('<br />\n')
-            html.write('<a href="{}">Previous Central Galaxy ({})</a>\n'.format(prevhtmlgalaxydir1, prevgalaxy[ii]))
-            html.write('<br />\n')
-
-            html.write('<br /><b><i>Last updated {}</b></i>\n'.format(js))
-            html.write('<br />\n')
-            html.write('</html></body>\n')
-            html.close()
-
-    # Make the plots.
-    if makeplots:
-        err = make_plots(sample, datadir=datadir, htmldir=htmldir, refband=refband,
-                         bands=bands, pixscale=pixscale, zcolumn=zcolumn, survey=survey,
-                         clobber=clobber, verbose=verbose, nproc=nproc, ccdqa=ccdqa,
-                         maketrends=maketrends)
-
-    try:
-        cmd = '/usr/bin/chgrp -R cosmo {}'.format(htmldir)
-        print(cmd)
-        err1 = subprocess.call(cmd.split())
-
-        cmd = 'find {} -type d -exec chmod 775 {{}} +'.format(htmldir)
-        print(cmd)
-        err2 = subprocess.call(cmd.split())
-
-        cmd = 'find {} -type f -exec chmod 664 {{}} +'.format(htmldir)
-        print(cmd)
-        err3 = subprocess.call(cmd.split())
-    except:
-        pass
-
-    #if err1 != 0 or err2 != 0 or err3 != 0:
-    #    print('Something went wrong updating permissions; please check the logfile.')
-    #    return 0
-    
-    return 1
